+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 ### Fixed
+- add TargetGroup, GroupPlace to [Group-Information](https://github.com/5pm-HDH/churchtools-api/pull/64)
+- generate Registration-Link in [PublicGroup](https://github.com/5pm-HDH/churchtools-api/pull/64)
 
 ## [1.3.0] - 2022-02-11